Better test coverage for Bcsec::Modes::HttpBasic.
The uncovered code was in HttpBasic#credentials; making it public makes
it easier to write specs for HttpBasic#credentials.

It also forced HttpBasic#credentials to deal with the possibility of an
Authorization header not being present  -- being public, it can be
called by any HttpBasic user, from anywhere, at any time.  Dealing with
this made it possible to implement #valid? in terms of #credentials.