Clear session after noninteractive request.  #2757.
The implementation in this commit makes the HTTP Basic scenario pass.  (I'm
still trying to develop a readable CAS proxy scenario.)

However, I'm not sure if this would be better expressed as another piece of
Rack middleware.  The current method is definitely shorter, but not very clear,
and it violates the declared Rack middleware order.

This also causes a few failures in Bcsec::Rack::Setup's spec, and
although they're not catastrophic failures, fixing them still
necessitates adding more method stubs, which feels dirty.  Maybe an
extra piece of middleware is the way to go.

Will investigate further.