Bring request parameter setup into line with Rack 1.3.
In Rack 1.3.0, Rack::Request#params does this:

    @params ||= GET.merge(POST)

It used to do this:

    GET.update(POST)

where @GET@ and @POST@ were built from data in @env@.  The result was
that Rack::Request#[]= in Rack < 1.3.0 modified the environment passed
into the request.

That doesn't happen anymore, so we need a different way of setting
parameters.