Increase session timeout on the session extension scenario.  #5156.
This gives an extra second of leeway for making requests, which
minimizes the occurrence of false failures.  I found it to be useful
when running bcsec's Cucumber tests on a 1.4 GHz Macbook Air.

Of course, it'd be nice to not have to mess with these sorts of timings,
but no other testing mechanism comes to mind.