When updating bcsec configuration, restart all spawned servers.  #5156.
The CAS scenarios spin up multiple servers (application, CAS server, and
proxy callback).  It's true that the application server is the only one
that requires a restart, but for now it's more convenient to just
restart them all.  Should a server end up being quite slow to start up
then we'll revisit this.