Reintroduce resolver hacks.
Originally, we threw in an otherwise unnecessary version constraint on
builder for ActiveRecord 3.0 to eliminate a dependency resolution run
that went out of control.   For more information, see:

    commit a159716ae6131fe3b39d94500f129c8cc976f03a
    Author: David Yip <yipdw@northwestern.edu>
    Date:   Tue Mar 1 14:15:36 2011 -0600

The hacks were removed once ActiveRecord was eliminated as a dependency:

    commit 6341023f0ffd158297b68bbe381ef0aa9dddcafa
    Author: Rhett Sutphin <rhett@detailedbalance.net>
    Date:   Sun Jul 17 02:50:34 2011 -0500

However, reintroducing ActiveRecord >= 2.3.12 (via RubyCAS-Server) seems
to have also reintroduced the dependency resolution problem.  Back to
the hacks we go.