Change strategy for avoiding infinite resolver sessions in bundler.
There are were two things which could each separately cause indefinite
resolves with bundler in this Gemfile. One, which was previously worked around
with the removed resolver_hacks group, was an oscillation between builder
2.1.2 and 3.x. This problem is solved by explicitly only working with later
versions in the 3.0 line of the Rails active* libraries, though other than
the reduction in search space I'm not sure why this fixes it.

The second was caused by the open reference to activerecord in the development
group. Bundler would spend a lot of time trying to find a version of AR that
was compatible with any given selected version of AS as well as all the
transitive dependencies to AR. This is now worked around by explicitly
depending on AR with the same version range as AS, since their actual version
numbers will always be the same.

It will remain a mystery why these problems only occur with the active*
libraries in the 3.0 line and not in 2.3 or 3.1.