Set SpawnedHttpServer pwd to tmpdir.
If a SpawnedHttpServer generates temporary files, there are three ways
to get a path:

1. Working directory
2. Accessing tmpdir
3. Making up a temporary directory

(3) is not acceptable: we want all test data to stay in the Aker
integrated tests scratch directory.  (2) is possible, but requires more
communication between the test process and the test server object.  (1)
requires a bit of care, but piggybacks on existing UNIX process
semantics and for the most part Just Works.

I'm going with (1).